#97 refactor: 히스토리 페이지 UI 폴더 1depth로 변경
