+++ src/pages/history/ui/history-completed-main-quests.module.scss
