+++ src/pages/history/ui/history-empty.module.scss
